+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.15](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.14...@synapsecns/sdk-router@0.2.15) (2023-09-05)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.2.14](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.13...@synapsecns/sdk-router@0.2.14) (2023-08-28)
 
 **Note:** Version bump only for package @synapsecns/sdk-router